moved all of Header's components into the header file as functions
