--- src/components/NavButtons.tsx
@@ -1,11 +1,0 @@
-import "../styles/NavButtons.scss";
-
-const NavButtons: React.FC = () => {
-  return (
-    <nav>
-      <button>Restaurants</button>
-      <button>Chefs</button>
-    </nav>
-  );
-};
-export default NavButtons;
