--- src/components/SearchSection.tsx
@@ -1,27 +1,0 @@
-import { useState } from "react";
-import { searchIcon } from "../assets/Icons";
-import "../styles/SearchSection.scss";
-const SearchSection: React.FC = () => {
-  const [showSearch, setShowSearch] = useState(false);
-  const toggleSearch = () => {
-    setShowSearch((prev) => !prev);
-  };
-  return (
-    <div className="search">
-      {showSearch && (
-        <input
-          className="search-box"
-          type="text"
-          placeholder="|Search for restaurant, cuisine, chef"
-        />
-      )}
-      <img
-        src={searchIcon}
-        alt="search-icon"
-        className="search-icon"
-        onClick={toggleSearch}
-      />
-    </div>
-  );
-};
-export default SearchSection;
