--- src/components/ShoppingcartToggle.tsx
@@ -1,5 +1,0 @@
-import { bagIcon } from "../assets/Icons";
-const ShoppingcartToggle: React.FC = () => {
-  return <img src={bagIcon} alt="bag-icon" />;
-};
-export default ShoppingcartToggle;
