--- src/components/UserToggle.tsx
@@ -1,5 +1,0 @@
-import { userIcon } from "../assets/Icons";
-const UserToggle: React.FC = () => {
-  return <img src={userIcon} alt="user-icon" />;
-};
-export default UserToggle;
