--- src/styles/LogoSection.scss
@@ -1,10 +1,0 @@
-div{
-    display: flex;
-    align-items: center;
-    span{
-        font-family: Helvetica;
-        font-weight: 200;
-        font-size: 1.8vw;
-        letter-spacing: 0.1vw;
-    }
-}
