--- src/styles/NavButtons.scss
@@ -1,13 +1,0 @@
-nav {
-  display: flex;
-  align-items: center;
-  gap: 0.1vw;
-}
-button {
-  background-color: white;
-  font-family: Helvetica;
-  font-weight: 200;
-  font-size: 1.4vw;
-  line-height: 100%;
-  letter-spacing: 0.2vw;
-}
