--- src/styles/SearchSection.scss
@@ -1,16 +1,0 @@
-.search{
-  display: flex;
-  align-items: center;
-  flex-direction: row-reverse;
-  position: absolute;
-  img {
-    position: absolute;
-    margin-right: 0.8vw;
-  }
-  input {
-    position: absolute;
-    width: 25vw;
-    height: 4vh;
-    margin-right: 0.6vw;
-  }
-}
